pyproject.toml
[project]
name = "quant-trading-bot"
version = "0.1.0"
description = "Three-candle Pivot Breakout Algorithmic Trading Bot"
authors = [
    {name = "User", email = "user@example.com"}
]
dependencies = [
    "pandas",
    "numpy",
    "matplotlib",
    "websockets",
    "python-dotenv",
    "pyyaml",
    "requests",
    "upstox-python-sdk",
    "urllib3==1.26.15",
    "protobuf"
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
